♻️ Split environment values by domain
