--- Sources/Luminare/Utilities/Extensions/EnvironmentValues+Extensions.swift
@@ -1,128 +1,0 @@
-//
-//  EnvironmentValues+Extensions.swift
-//  Luminare
-//
-//  Created by Kai Azim on 2024-04-05.
-//
-
-import SwiftUI
-
-// MARK: - Internal
-
-extension EnvironmentValues {
-    @Entry var luminareClickedOutside: Bool = false
-}
-
-// MARK: - Common
-
-public extension EnvironmentValues {
-    /// Currently, it is impossible to read the `.tint()` modifier on a view
-    /// this is a custom environement value as an alternative implementation of it
-    /// in practice, it should always be synchronized with `.tint()`
-    @Entry var luminareTintColor: Color = .accentColor
-
-    @Entry var luminareAnimation: Animation = .smooth(duration: 0.2)
-    @Entry var luminareAnimationFast: Animation = .easeInOut(duration: 0.1)
-
-    @Entry var luminareDismiss: () -> () = {}
-}
-
-// MARK: - Windows
-
-public extension EnvironmentValues {
-    // MARK: Modal
-
-    @Entry var luminareModalStyle: LuminareModalStyle = .sheet
-    @Entry var luminareModalCornerRadius: CGFloat = 12
-    @Entry var luminareModalPresentation: LuminareModalPresentation = .windowCenter
-    @Entry var luminareModalClosesOnDefocus: Bool = false
-    @Entry var luminareIsInsideModal: Bool = false
-
-    // MARK: Color Picker
-
-    @Entry var luminareColorPickerHasCancel: Bool = false
-    @Entry var luminareColorPickerHasDone: Bool = false
-}
-
-// MARK: - View
-
-public extension EnvironmentValues {
-    @Entry var luminareCornerRadii: RectangleCornerRadii = .init(12)
-    @Entry var luminareMinHeight: CGFloat = 32
-
-    @Entry var luminareBorderedStates: LuminareBorderStates = .default
-    @Entry var luminareFilledStates: LuminareFillStates = .default
-    @Entry var luminareSurfaceStyle: LuminareSurfaceStyle = .plateau
-    @Entry var luminareHasDividers: Bool = true
-
-    @Entry var luminareContentMarginsTop: CGFloat = 0
-    @Entry var luminareContentMarginsLeading: CGFloat = 0
-    @Entry var luminareContentMarginsBottom: CGFloat = 0
-    @Entry var luminareContentMarginsTrailing: CGFloat = 0
-
-    // MARK: Form
-
-    @Entry var luminareFormSpacing: CGFloat = 16
-    @Entry var luminareFormLayout: LuminareFormLayout = .stacked
-
-    /// A naming convention: `titleBar` for SwiftUI, `titlebar` for AppKit, and `title bar` for natural language
-    @Entry var luminareTitleBarHeight: CGFloat = 50
-
-    // MARK: Section
-
-    @Entry var luminareSectionLayout: LuminareSectionLayout = .stacked
-    @Entry var luminareSectionHorizontalPadding: CGFloat = 8
-    @Entry var luminareIsInsideSection: Bool = false
-    @Entry var luminareTopLeadingRounded: Bool = true
-    @Entry var luminareTopTrailingRounded: Bool = true
-    @Entry var luminareBottomLeadingRounded: Bool = true
-    @Entry var luminareBottomTrailingRounded: Bool = true
-
-    /// If 0, then luminareSection will be of fixed size.
-    @Entry var luminareSectionMaxWidth: CGFloat? = .infinity
-
-    // MARK: Button Row
-
-    @Entry var luminareButtonRowSpacing: CGFloat = 4
-
-    // MARK: Compose
-
-    @Entry var luminareComposeControlSize: LuminareComposeControlSize = .automatic
-
-    // MARK: Tool Tip
-
-    @Entry var luminareToolTipTrigger: LuminareToolTipTrigger = .hover
-    @Entry var luminareToolTipShade: LuminareToolTipShade = .styled
-
-    // MARK: Stepper
-
-    @available(macOS 15.0, *)
-    @Entry var luminareStepperAlignment: LuminareStepperAlignment = .trailing
-    @available(macOS 15.0, *)
-    @Entry var luminareStepperDirection: LuminareStepperDirection = .horizontal
-
-    // MARK: Compact Picker
-
-    @Entry var luminareCompactPickerStyle: LuminareCompactPickerStyle = .menu
-
-    // MARK: List
-
-    @Entry var luminareListItemCornerRadii: RectangleCornerRadii = .init(2)
-
-    @Entry var luminareListItemHeight: CGFloat = 50
-    @Entry var luminareListItemHighlightOnHover: Bool = false
-    @Entry var luminareItemBeingHovered: Bool = false
-    @Entry var luminareListFixedHeightUntil: CGFloat? = nil
-
-    // MARK: Sidebar
-
-    @Entry var luminareSidebarOverflow: CGFloat = 50
-
-    // MARK: Slider
-
-    @Entry var luminareSliderLayout: LuminareSliderLayout = .regular
-
-    // MARK: Slider Picker
-
-    @Entry var luminareSliderPickerLayout: LuminareSliderPickerLayout = .regular
-}
